don't touch default min sdk version
target `all_jdk_libs` will call [r8 to shrink the code-size][1] and use
`default_min_sdk_version` as value of argument of "--min-api". In this
version r8 fails when min api >=26.

So we should not set argument of `default_min_sdk_version`, instead we
can directly set argument of `android64_ndk_api_level`, which gn will
use to set the cflag of "--target".

Setting argument of `default_min_sdk_version` also can change the
default value of `android64_ndk_api_level`, but now seems it is not the
correct place because it will impact other logic, such as r8.

Test: make ninja T=build/android:all_jdk_libs

[1]:https://nphausg.medium.com/android-journey-proguard-d8-r8-what-are-they-e8f2bfe079a7

Signed-off-by: Wang Chen <wangchen20@iscas.ac.cn>